+++ app/models.py
@@ -17,8 +17,8 @@ from sqlalchemy import (
 )
 
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from db import Base
-from constants import Currency
+from config.db import Base
+from config.constants import Currency
 
 class Account(Base):
     __tablename__ = "accounts"
