+++ app/templates/config.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8" />
   <title>Configuración</title>
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
-  <link rel="stylesheet" href="../css/layout.css">
+  <link rel="stylesheet" href="/static/css/layout.css">
 </head>
 <body class="d-flex flex-column min-vh-100">
   <header class="navbar navbar-light bg-light p-2">
@@ -70,6 +70,6 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
-  <script type="module" src="../js/config.js"></script>
+  <script type="module" src="/static/js/config.js"></script>
 </body>
 </html>
