+++ app/templates/index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8" />
   <title>Movimientos</title>
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
-  <link rel="stylesheet" href="css/layout.css">
+  <link rel="stylesheet" href="/static/css/layout.css">
 </head>
 <body class="d-flex flex-column min-vh-100">
   <header class="navbar navbar-light bg-light p-2">
@@ -23,7 +23,7 @@
     </div>
     <div class="offcanvas-body">
       <ul class="list-unstyled">
-        <li><a href="/config" class="text-decoration-none">Configuración</a></li>
+        <li><a href="/config.html" class="text-decoration-none">Configuración</a></li>
       </ul>
     </div>
   </div>
@@ -82,6 +82,6 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
-  <script type="module" src="js/main.js"></script>
+  <script type="module" src="/static/js/main.js"></script>
 </body>
 </html>
