+++ docker-compose.yml
@@ -2,7 +2,7 @@ services:
   app:
     container_name: movdin-app
     hostname: movdin-app
-    build: ./app
+    build: .
     environment:
       DATABASE_URL: ${DATABASE_URL:-postgresql+psycopg2://postgres:postgres@db_TA:5432/db_TA}
     expose:
